requirements.txt
# Core
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3

# NLP
spacy==3.7.2
python-dotenv==1.0.0

# Vector DB
chromadb==0.4.22
sentence-transformers==2.3.1

# LLM - UPDATED
google-generativeai==0.4.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
trafilatura==1.6.3
urllib3==2.1.0
selenium==4.15.2
webdriver-manager==4.0.1

# PDF - ENHANCED
pdfplumber==0.10.3
PyPDF2==3.0.1
camelot-py[cv]==0.11.0
tabula-py==2.8.2

# Utilities
tqdm==4.66.1